Добавляет header, footer, catalog. Пробрасывает onMovieCardClick
