+++ src/components/footer/footer.test.js
@@ -1,0 +1,12 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import Footer from "./footer";
+
+it(`Should Footer render correctly`, () => {
+  const tree = renderer
+    .create(
+        <Footer/>
+    ).toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
