+++ src/components/main/main.jsx
@@ -1,39 +1,26 @@
 import React from "react";
 import PropTypes from "prop-types";
-import MovieList from "../movie-list/movie-list.jsx";
-import {filterMoviesByGenre, getGenres} from "../../utils/common/common";
-import GenreList from "../genre-list/genre-list.jsx";
-import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer";
-import {ShowedMovies} from "../../const";
+import Header from "../header/header.jsx";
+import Footer from "../footer/footer.jsx";
+import Catalog from "../catalog/catalog.jsx";
 
 const Main = (props) => {
-  const {promo, movies, genres, activeGenre, onMovieCardClick, onGenreClick} = props;
+  const {promo, onMovieCardClick} = props;
   const {title, genre, year} = promo;
 
-  // TODO: Вынести шапку и подвал в отдельный компонент.
-  //  NOTE: шапка различается на страницах
   return (
     <>
       <section className="movie-card">
+        <h1 className="visually-hidden">WTW</h1>
+        <Header
+          isMainPage={true}
+        >
+        </Header>
+
         <div className="movie-card__bg">
           <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel"/>
         </div>
-        <h1 className="visually-hidden">WTW</h1>
-        <header className="page-header movie-card__head">
-          <div className="logo">
-            <a className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-          <div className="user-block">
-            <div className="user-block__avatar">
-              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63"/>
-            </div>
-          </div>
-        </header>
+
         <div className="movie-card__wrap">
           <div className="movie-card__info">
             <div className="movie-card__poster">
@@ -66,38 +53,11 @@ const Main = (props) => {
       </section>
 
       <div className="page-content">
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
+        <Catalog
+          onMovieCardClick={onMovieCardClick}
+        />
 
-          <GenreList
-            genres={genres}
-            activeGenre={activeGenre}
-            onClick={onGenreClick}
-          />
-
-          <MovieList
-            movies={movies}
-            onMovieCardClick={onMovieCardClick}
-          />
-
-          <div className="catalog__more">
-            <button className="catalog__button" type="button">Show more</button>
-          </div>
-        </section>
-
-        <footer className="page-footer">
-          <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer/>
       </div>
     </>
   );
@@ -109,35 +69,7 @@ Main.propTypes = {
     genre: PropTypes.string.isRequired,
     year: PropTypes.number.isRequired,
   }),
-  movies: PropTypes.arrayOf(
-      PropTypes.shape({
-        title: PropTypes.string.isRequired,
-        genre: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-        year: PropTypes.number.isRequired,
-        rating: PropTypes.number.isRequired,
-        votes: PropTypes.number.isRequired,
-        producer: PropTypes.string.isRequired,
-        actors: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-        description: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-      })
-  ).isRequired,
-  genres: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-  activeGenre: PropTypes.string.isRequired,
-  onGenreClick: PropTypes.func.isRequired,
   onMovieCardClick: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  movies: filterMoviesByGenre(state.movies, state.genre).slice(0, ShowedMovies.ON_START),
-  activeGenre: state.genre,
-  genres: getGenres(state.movies),
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  onGenreClick(genre) {
-    dispatch(ActionCreator.setFilterByGenre(genre));
-  },
-});
-
-export {Main};
-export default connect(mapStateToProps, mapDispatchToProps)(Main);
+export default Main;
